dataset/requirements.txt
beautifulsoup4==4.11.2
Markdown==3.4.1
datasets==2.9.0
PyYAML==6.0
transformers @ https://github.com/huggingface/transformers/archive/32ff06403d48fa018a33ccab02a5f9d3eb8b078b.zip
torch==2.0.0
sentencepiece==0.1.97
openpyxl==3.1.2
pdf2image==1.16.3
pdfCropMargins==1.1.18
